dev/requirements.txt
django
django-allauth
django-composed-configuration
django-configurations
django-debug-toolbar
django-oauth-toolkit==1.3.2
django-tailwind
rich
